Consolidate retention and proof modules
