--- src/retnet_retention.py
@@ -1,3 +1,0 @@
-from .retention import RetNetRetention
-
-__all__ = ["RetNetRetention"]
